Created Health UI, Refactored DamageDealer
See _Docs/TODO.txt for how the DamageDealer should be refactored.

+++ SimpleDungeonCrawler_URP/Assets/Project/_Scripts/Damage/Core/IDamagable.cs
@@ -8,6 +8,6 @@ namespace Project.Damage
 {
     public interface IDamagable
 	{
-		void TakeDamage(int _damageAmount);
+		void TakeDamage(float _damageAmount);
 	}
 }
